requirements.txt
pymysql
djangorestframework
markdown
django-filter